refactor: normalize folder casing
