+++ src/App.js
@@ -1,10 +1,10 @@
 import { Routes, Route } from "react-router-dom";
-import Home from "./pages/Home";
-import Items from "./pages/Items/Items.js";
+import Home from "./pages/home/Home.js";
+import Items from "./pages/items/Items.js";
 import DefaultLayout from "./layouts/DefaultLayout";
 import "pretendard/dist/web/static/pretendard.css";
 import "./styles/global.css";
-import Additem from "./pages/Additem";
+import Additem from "./pages/additem/Additem.js";
 
 function App() {
