+++ src/layouts/DefaultLayout.js
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router-dom";
-import Navbar from "../components/Navbar/Navbar.js";
+import Navbar from "../components/navbar/Navbar.js";
 import styles from "./DefaultLayout.module.css"
